backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
openai>=1.50.0
python-dotenv>=1.0.0
redis>=5.2.0
pydantic>=2.10.0
